Add schema manager
